requirements.txt
numpy==1.15.3
Pillow==5.3.0
pkg-resources==0.0.0
six==1.11.0
torch==0.4.1
torchvision==0.2.1
imageio
scipy
segmentation-models-pytorch
BeautifulSoup4
lxml
wandb
albumentations
